fix unit test

+++ src/plugins/intel_gpu/tests/unit/test_cases/gated_delta_net.cpp
@@ -85,12 +85,12 @@ struct gated_delta_net_gpu_test : public ::testing::TestWithParam<gated_delta_ne
     void l2norm(float* a, size_t n) {
         float eps = 0.000001;
         float sum = 0.0f;
-        for (int j = 0; j < n; j++) {
+        for (size_t j = 0; j < n; j++) {
             sum += a[j] * a[j];
         }
         sum += eps;
         sum = 1 / sqrt(sum);
-        for (int j = 0; j < n; j++) {
+        for (size_t j = 0; j < n; j++) {
             a[j] = a[j] * sum;
         }
     }
@@ -130,7 +130,7 @@ struct gated_delta_net_gpu_test : public ::testing::TestWithParam<gated_delta_ne
                         float b_g = g[i_b * G_B_STRIDE + i * this->H + i_h];
                         float b_beta = beta[i_b * G_B_STRIDE + i * this->H + i_h];
                         b_g = exp(b_g);
-                        for (int j = 0; j < this->K; j++) {
+                        for (size_t j = 0; j < this->K; j++) {
                             b_k[j] = k_ptr[i * this->K * this->HK + j];
                             b_q[j] = q_ptr[i * this->K * this->HK + j];
                         }
